Refactored PubSubPublisherTemplate and PubSubSubscriberTemplate into their own respective packages
